Fix localization in Manage panel
